购物车价格相关字段修改为浮点型数据

+++ wstmart/common/model/Carts.php
@@ -290,7 +290,7 @@ class Carts extends Base{
 			}
 			//未选中商品 已下架商品  库存小于或等于0商品不予统计金额
 			if ($v['isCheck'] == 1 && $v['isSale'] !=0 && $v['goodsStock'] > 0 && $v['specStock'] > 0) {
-				$carts[$v['shopId']]['goodsMoney'] = bcdiv($carts[$v['shopId']]['goodsMoney'] + $v['shopPrice'] * $v['cartNum'],1, 2);
+				$carts[$v['shopId']]['goodsMoney'] = (float)bcdiv($carts[$v['shopId']]['goodsMoney'] + $v['shopPrice'] * $v['cartNum'],1, 2);
 				$goodsTotalMoney = $goodsTotalMoney + $v['shopPrice'] * $v['cartNum'];
 				$goodsTotalNum+=$v['cartNum'];
 			}
@@ -349,7 +349,7 @@ class Carts extends Base{
 			if(!isset($v['list']))unset($carts[$key]);
 		}
 	
-		$cartData = ['carts' => array_values($carts), 'goodsTotalMoney' =>  bcdiv($goodsTotalMoney,1, 2), 'goodsTotalNum' => (int)$goodsTotalNum, 'promotionMoney' => '0.00','canNotBuy' => $canNotBuy];
+		$cartData = ['carts' => array_values($carts), 'goodsTotalMoney' => (float)bcdiv($goodsTotalMoney,1, 2), 'goodsTotalNum' => (int)$goodsTotalNum, 'promotionMoney' => '0.00','canNotBuy' => $canNotBuy];
 		//店铺优惠活动监听
 		//hook("afterQueryCarts",["carts"=>&$cartData,'isSettlement'=>$isSettlement,'isVirtual'=>false,'uId'=>$userId]);
 		return $cartData;
